fixed bitmap offset being a float
getBoundingClientRect() returns float numbers and therefore the bitmap offset, which uses this.size, which is derived from it, can become a float in some cases.
This results in breaking the whole canvas bitmap drawing (i.e. the color circle disappears), as arrays expect integers as indexers.
Using  Math.floor before calculating the size dimensions prevents this.