Use `attached` rather than `ready` for initial color
Binding engines such as Aurelia don’t have the value provided to the
element at `ready` time. Using `attached` allows binding engines a
chance to provide an initial value.

This also moves the setting of `immediateColor` to after the “listen
once” event subscription so that it will be guaranteed to fire, even on
fast hardware.